[refactor] #13 책임 분리를 위한 store와 map 분리
