+++ src/main/java/org/netway/dongnehankki/map/application/MapBoundaryCalculator.java
@@ -1,4 +1,4 @@
-package org.netway.dongnehankki.store.application;
+package org.netway.dongnehankki.map.application;
 
 import org.springframework.stereotype.Component;
