+++ src/main/java/org/netway/dongnehankki/map/application/MapService.java
@@ -1,10 +1,10 @@
-package org.netway.dongnehankki.store.application;
+package org.netway.dongnehankki.map.application;
 
 import java.util.List;
 
 import org.netway.dongnehankki.store.domain.Store;
-import org.netway.dongnehankki.store.dto.request.MapRequest;
-import org.netway.dongnehankki.store.dto.response.MapResponse;
+import org.netway.dongnehankki.map.dto.request.MapRequest;
+import org.netway.dongnehankki.map.dto.response.MapResponse;
 import org.netway.dongnehankki.store.infrastructure.StoreRepository;
 import org.springframework.stereotype.Service;
